# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: MK Digital Operations Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /security/vulnerabilities:
    get:
      operationId: listVulnerabilities
      tags: [security]
      summary: List vulnerabilities
      parameters:
        - name: severity
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Vulnerabilities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedVulnerabilityList"
    post:
      operationId: createVulnerability
      tags: [security]
      summary: Create vulnerability record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VulnerabilityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vulnerability"
  /security/vulnerabilities/{id}:
    patch:
      operationId: updateVulnerability
      tags: [security]
      summary: Update vulnerability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VulnerabilityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vulnerability"
    delete:
      operationId: deleteVulnerability
      tags: [security]
      summary: Delete vulnerability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /security/vulnerabilities/{id}/restore:
    post:
      operationId: restoreVulnerability
      tags: [security]
      summary: Restore a soft-deleted vulnerability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored vulnerability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vulnerability"
        "404":
          description: Not found
...
  /software:
    get:
      operationId: listSoftware
      tags: [software]
      summary: List software inventory
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: endOfLife
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Software list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSoftwareList"
    post:
      operationId: createSoftware
      tags: [software]
      summary: Add software to inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoftwareInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoftwareItem"
  /software/{id}:
    patch:
      operationId: updateSoftware
      tags: [software]
      summary: Update software record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoftwareUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoftwareItem"
    delete:
      operationId: deleteSoftware
      tags: [software]
      summary: Delete software record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /software/{id}/restore:
    post:
      operationId: restoreSoftware
      tags: [software]
      summary: Restore a soft-deleted software record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored software record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoftwareItem"
        "404":
          description: Not found

  # ─── Documentation ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Vulnerability:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        severity:
          type: string
        status:
          type: string
        applicationId:
          type: ["integer", "null"]
        applicationName:
          type: ["string", "null"]
        cveId:
          type: ["string", "null"]
        affectedComponent:
          type: ["string", "null"]
        version:
          type: ["string", "null"]
        vendor:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        installationDate:
          type: ["string", "null"]
        licenseType:
          type: ["string", "null"]
        licenseExpiration:
          type: ["string", "null"]
        endOfLifeDate:
          type: ["string", "null"]
        discoveredAt:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type: ["string", "null"]
      required: [id, title, severity, status, createdAt, updatedAt]
    VulnerabilityInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        severity:
          type: string
        status:
          type: string
        applicationId:
          type: integer
        cveId:
          type: string
        affectedComponent:
          type: string
        version:
          type: string
        vendor:
          type: string
        category:
          type: string
        installationDate:
          type: string
        licenseType:
          type: string
        licenseExpiration:
          type: string
        endOfLifeDate:
          type: string
        discoveredAt:
          type: string
        resolvedAt:
          type: string
        assignedTo:
          type: string
        notes:
          type: string
        ownerId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
      required: [title, severity, status]
    VulnerabilityUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        severity:
          type: string
        status:
          type: string
        applicationId:
          type: integer
        cveId:
          type: string
        affectedComponent:
          type: string
        version:
          type: string
        vendor:
          type: string
        category:
          type: string
        installationDate:
          type: string
        licenseType:
          type: string
        licenseExpiration:
          type: string
        endOfLifeDate:
          type: string
        discoveredAt:
          type: string
        resolvedAt:
          type: string
        assignedTo:
          type: string
        notes:
          type: string
        ownerId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
...
    SoftwareItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        installedVersion:
          type: ["string", "null"]
        latestVersion:
          type: ["string", "null"]
        vendor:
          type: ["string", "null"]
        license:
          type: ["string", "null"]
        supported:
          type: boolean
        endOfLife:
          type: boolean
        endOfLifeDate:
          type: ["string", "null"]
        upgradeAvailable:
          type: boolean
        applicationId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type: ["string", "null"]
      required: [id, name, type, supported, endOfLife, upgradeAvailable, createdAt, updatedAt]
    SoftwareInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
        installedVersion:
          type: string
        latestVersion:
          type: string
        vendor:
          type: string
        license:
          type: string
        supported:
          type: boolean
        endOfLife:
          type: boolean
        endOfLifeDate:
          type: string
        upgradeAvailable:
          type: boolean
        applicationId:
          type: integer
        notes:
          type: string
        ownerId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
      required: [name, type]
    SoftwareUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        installedVersion:
          type: string
        latestVersion:
          type: string
        vendor:
          type: string
        license:
          type: string
        supported:
          type: boolean
        endOfLife:
          type: boolean
        endOfLifeDate:
          type: string
        upgradeAvailable:
          type: boolean
        applicationId:
          type: integer
        notes:
          type: string
        ownerId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]

    # ── Documentation ─────────────────────────────────────────────────────
...
    PaginatedVulnerabilityList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Vulnerability"
        total:
          type: integer
...
    PaginatedSoftwareList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SoftwareItem"
        total:
          type: integer
...