# lib/api-zod/src/generated/types/softwareItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MK Digital Operations Center API
 * OpenAPI spec version: 1.0.0
 */

export interface SoftwareItem {
  id: number;
  name: string;
  type: string;
  /** @nullable */
  installedVersion?: string | null;
  /** @nullable */
  latestVersion?: string | null;
  /** @nullable */
  vendor?: string | null;
  /** @nullable */
  license?: string | null;
  supported: boolean;
  endOfLife: boolean;
  /** @nullable */
  endOfLifeDate?: string | null;
  upgradeAvailable: boolean;
  /** @nullable */
  applicationId?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  ownerId?: number | null;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  teamId?: number | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  deletedAt?: string | null;
}